+++ packages/cli/src/commands/dev.ts
@@ -1,24 +1,10 @@
-import { GluegunCommand, GluegunParameters } from 'gluegun';
+import { GluegunCommand } from 'gluegun';
 import { getMitosisConfig } from '../helpers/get-mitosis-config';
 import { build, clean } from '../build/build';
 import chokidar from 'chokidar';
 import { dev } from '../build/dev';
-import { MitosisConfig, Target } from '@builder.io/mitosis';
-
-export const getTargets = (mitosisConfig: MitosisConfig, cliOpts: GluegunParameters['options']) => {
-  const targetsFromCli: Target[] = (cliOpts.targets || '').split(',');
-  const excludeTargetsMap: Record<Target, true> = (cliOpts.excludeTargets || '')
-    .split(',')
-    .reduce((accu, t) => {
-      accu[t] = true;
-      return accu;
-    }, {});
-
-  const targets = Array.from(new Set([...mitosisConfig.targets, ...targetsFromCli])).filter(
-    (t) => t && !excludeTargetsMap[t],
-  );
-  return targets;
-};
+import { getTargets } from '../helpers/get-targets';
+import { Target } from '@builder.io/mitosis';
 
 const command: GluegunCommand = {
   name: 'dev',
@@ -39,27 +25,31 @@ const command: GluegunCommand = {
     } else {
       files.push(...config.files);
     }
+
+    console.log('mitosis dev server started. running build command...');
+
     await build({
       ...config,
       targets,
     });
 
     const watcher = chokidar.watch(files, {}).on('change', async (path) => {
-      performance.mark('start');
+      const start = new Date();
       console.log(`detected change in ${path}, rebuilding...`);
       try {
-        await dev({ config: { ...config, targets }, paths: [path] });
-        console.log(`Rebuilt in ${performance.measure('start').duration}`);
+        const data = await dev({ config: { ...config, targets }, paths: [path] });
+        const end = new Date();
+        const time = (end.getTime() - start.getTime()) / 1000;
+        log({ data, time });
       } catch (e) {
-        console.log('error rebuilding, save the file again to try again');
+        console.log('error rebuilding');
       }
     });
 
     const overrideWatcher = chokidar
       .watch(config.overridesDir || 'overrides', {})
-      .on('change', async (path) => {
-        console.log(`detected change in override folder: ${path}.
-       Rerunning build (this might take a while)`);
+      .on('change', async () => {
+        console.log(`detected change in override folder. Rerunning build command...`);
 
         await clean({ ...config, targets });
 
@@ -74,6 +64,22 @@ const command: GluegunCommand = {
       overrideWatcher.close();
     });
   },
+};
+
+export const log = ({
+  data,
+  time,
+}: {
+  data: { target: Target; components: number; files: number }[];
+  time: number;
+}) => {
+  const components = data.map((a) => a.components).reduce((a, b) => a + b, 0);
+  const files = data.map((a) => a.files).reduce((a, b) => a + b, 0);
+  const targets = data.map((a) => a.target).length;
+
+  console.log(
+    `Generated ${files} files and ${components} components in ${targets} languages in ${time}s`,
+  );
 };
 
 module.exports = command;
