+++ packages/cli/src/helpers/get-targets.ts
@@ -1,0 +1,17 @@
+import { MitosisConfig, Target } from "@builder.io/mitosis";
+import { GluegunParameters } from "gluegun";
+
+export const getTargets = (mitosisConfig: MitosisConfig, cliOpts: GluegunParameters['options']) => {
+  const targetsFromCli: Target[] = (cliOpts.targets || '').split(',');
+  const excludeTargetsMap: Record<Target, true> = (cliOpts.excludeTargets || '')
+    .split(',')
+    .reduce((accu, t) => {
+      accu[t] = true;
+      return accu;
+    }, {});
+
+  const targets = Array.from(new Set([...mitosisConfig.targets, ...targetsFromCli])).filter(
+    (t) => t && !excludeTargetsMap[t],
+  );
+  return targets;
+};
